requirements.txt
# RAG Dashboard System Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# AI and ML
openai==1.3.7
google-generativeai==0.3.2
tiktoken==0.5.1
numpy==1.24.3

# Vector Database
chromadb==0.4.15

# Document Processing
PyPDF2==3.0.1
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0

# Environment and Configuration
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Data Visualization (for Streamlit dashboard)
plotly==5.17.0
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0
flake8==6.1.0

# Additional Utilities
pathlib2==2.3.7
typing-extensions==4.8.0